+++ .gitignore
@@ -2,7 +2,6 @@
 # will have compiled files and executables
 debug/
 target/
-.vscode*
 
 # Remove Cargo.lock from gitignore if creating an executable, leave it for libraries
 # More information here https://doc.rust-lang.org/cargo/guide/cargo-toml-vs-cargo-lock.html
@@ -16,4 +15,4 @@ Cargo.lock
 
 .env*
 
-.vscode/
+.vscode/**/*
