+++ migration/src/m20220101_000001_create_table.rs
@@ -9,23 +9,23 @@ impl MigrationTrait for Migration {
         // Replace the sample below with your own migration scripts
         todo!();
 
-        manager
-            .create_table(
-                Table::create()
-                    .table(Post::Table)
-                    .if_not_exists()
-                    .col(
-                        ColumnDef::new(Post::Id)
-                            .integer()
-                            .not_null()
-                            .auto_increment()
-                            .primary_key(),
-                    )
-                    .col(ColumnDef::new(Post::Title).string().not_null())
-                    .col(ColumnDef::new(Post::Text).string().not_null())
-                    .to_owned(),
-            )
-            .await
+        // manager
+        //     .create_table(
+        //         Table::create()
+        //             .table(Post::Table)
+        //             .if_not_exists()
+        //             .col(
+        //                 ColumnDef::new(Post::Id)
+        //                     .integer()
+        //                     .not_null()
+        //                     .auto_increment()
+        //                     .primary_key(),
+        //             )
+        //             .col(ColumnDef::new(Post::Title).string().not_null())
+        //             .col(ColumnDef::new(Post::Text).string().not_null())
+        //             .to_owned(),
+        //     )
+        //     .await
     }
 
     async fn down(&self, manager: &SchemaManager) -> Result<(), DbErr> {
