+++ src/domain/repository/auth.rs
