+++ src/main.rs
@@ -1,6 +1,5 @@
 mod api;
 mod components;
-mod database;
 mod domain;
 mod util;
 
@@ -8,6 +7,8 @@ use axum::{routing::get, Router};
 use std::net::SocketAddr;
 use tower_http::trace::{self, TraceLayer};
 use tracing::Level;
+
+use crate::domain::repository;
 
 #[tokio::main]
 async fn main() {
@@ -18,8 +19,9 @@ async fn main() {
 
     tracing_subscriber::fmt().with_target(false).pretty().init();
 
-    let db = database::connect_to_db().await.unwrap();
-    let translator = util::localization::Translator::new("es_ES".to_string(), "locales");
+    let db = repository::connect_to_db().await.unwrap();
+    let translator =
+        util::localization::Translator::new("es_ES".to_string(), "./src/util/localization");
 
     let state = domain::AppState {
         supabase_api_key: std::env::var("SUPABASE_API_KEY")
